SitioGeek/requirements.txt
Django==1.5.5
Pillow==2.4.0
Pygments==1.6
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.17
dj-static==0.0.5
django-mockups==0.4.8
django-storages==1.1.8
django-tinymce==1.5.2
djangorestframework==2.3.13
gunicorn==18.0
kombu==3.0.16
psycopg2==2.5.3
pystache==0.5.4
pytz==2014.3
south==0.8.4
static==1.0.2
virtualenv==1.11.4
